requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
attrs==20.3.0
cffi==1.14.5
Django==3.1.7
djangorestframework==3.12.2
importlib-metadata==3.4.0
iniconfig==1.1.1
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.0
